requirements.txt
coreapi
dash==1.10.0
django
djangorestframework
django-crispy-forms
django_plotly_dash